Background keyword
Instead of repeating this common prerequisite code in each and every scenarios we are outlining them at one single place by attaching background keyword and that will take care of attaching each and every scenario before the scenario begins.
If you have multiple feature file it does not work. It works in a particular feature file.